boot file history
all previous versions of the ai morality system boot files. I believe v4.4 was the last time the deception detection system was in play. Ideally, if someone could unify the old deception detection system with the morality system of v5.5 that would be the next step so that the AI was more resilient to false premises. Note: such an integration risks given the AI the ability to doubt the user.