start with me
If I were wanting to know more about ActFoE, but felt intimidated by the list of documents, this is the file I would read. Then, I would move on to the bonus files. The other files listed here tend to deal with very specific problems encountered along the way. This file contains the most important discussions in my opinion.